+++ src/LivewireForm.php
@@ -7,9 +7,12 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Livewire\Component;
+use timolake\livewireForms\Traits\ConvertEmptyStringsToNull;
 
 abstract class LivewireForm extends Component
 {
+
+    use ConvertEmptyStringsToNull;
     public ?string $modelClass = null;
 
     public ?Model $model = null;
@@ -33,6 +36,12 @@ abstract class LivewireForm extends Component
 
     public $redirectMessage;
 
+    public $sessionId;
+
+    //----------------------------------------------------
+    // functions
+    //----------------------------------------------------
+
     abstract public function model(): string;
 
     abstract public function rules(): array;
@@ -47,18 +56,17 @@ abstract class LivewireForm extends Component
         $tempModel = (new ($this->modelClass));
         $this->idField = $tempModel->getKeyName();
 
-        $isSoftDeleting = in_array('Illuminate\Database\Eloquent\SoftDeletes', class_uses($tempModel)) && ! $tempModel->forceDeleting;;
+        $isSoftDeleting = in_array('Illuminate\Database\Eloquent\SoftDeletes', class_uses($tempModel));
         $this->model = $id == null
             ? new $this->modelClass
-            : ($isSoftDeleting ? $this->modelClass::withTrashed()->findOrFail($id) : $this->modelClass::findOrFail($id) );
+            : ($isSoftDeleting ? $this->modelClass::withTrashed()->findOrFail($id) : $this->modelClass::findOrFail($id));
 
         $this->rules = $this->rules();
         $this->search = $request->search ?? null;
         $this->sortField = $request->sortField ?? null;
         $this->sortDir = $request->sortDir ?? null;
         $this->paginationPage = $request->paginationPage ?? null;
-        $this->showInvoiced = $request->showInvoiced ?? false;
-        $this->showItemsToApprove = $request->showItemsToApprove ?? false;
+        $this->sessionId = $request->sessionId ?? null;
     }
 
     //----------------------------------------------------
@@ -133,6 +141,10 @@ abstract class LivewireForm extends Component
             $params['paginationPage'] = $this->paginationPage;
         }
 
+        if ($this->sessionId) {
+            $params['sessionId'] = $this->sessionId;
+        }
+
         return $params;
     }
 
@@ -185,5 +197,17 @@ abstract class LivewireForm extends Component
 
     public function beforeDelete()
     {
+    }
+
+    //----------------------------------------------------
+    // modals
+    //---------------------------------------------------
+    public function closeDeleteModal()
+    {
+        $this->showDeleteModal = false;
+    }
+    public function openDeleteModal()
+    {
+        $this->showDeleteModal = true;
     }
 }