+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
+
 All notable changes to `livewire-forms` will be documented in this file.
+## v1.6 - 2023-11-28
+* add delete modal functions
+## v1.5 - 2023-10-06
+* add sessionId to url params
+## v1.4 - 2023-09-17
+* convert empty strings to null
+## v1.3 - 2023-08-07
+* ItemForm: add validation attributes
+## v1.2.2 - 2023-08-03
+* ItemForm: fix merge
+## v1.2.1 - 2023-08-03
+* ItemForm: bugfix create
+## v1.2 - 2023-08-02
+* add ItemForm (form with sublist)
+## v1.1 - 2023-05-29
+* add laravel 10 support
+## v1.0.4 - 2023-02-17
+* bugfix for trashed
 ## v1.0.3 - 2023-01-20
 * add support for trashed
 ## v1.0.0 - 2022-08-10